Integration
- Add Python/pipenv dependencies
- Add rawtext endpoint to flask app
- Update README
