+++ README.md
@@ -1,7 +1,39 @@
 # tiktok-deepfake
-### Run
+## Install Dependencies
+Install Python3/pip and Node/npm.
+
+Install [pipenv](https://pipenv.pypa.io/en/latest/) Python package manager:
+
+```pip install --user pipenv```
+
+Enter the backend directory:
+
+```cd api```
+
+Install the python dependencies:
+
+```pipenv install```
+
+Install the node dependencies:
+
 ```
-python api/api.py
-cd app
+cd ../app
+npm install
+```
+
+## Run
+```
+cd api
+pipenv run flask run
+cd ../app
 npm start
 ```
+
+## Test
+Make a test request:
+
+```curl --header "Content-Type: application/json"   --request POST   --data '{"rawText": "hello world!"}'   http://localhost:5000/api/rawtext```
+
+## Auto-format the code
+
+```pipenv run black app.py```
