+++ api/Pipfile
@@ -1,0 +1,14 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flask = "*"
+flask-socketio = "*"
+
+[dev-packages]
+black = "*"
+
+[requires]
+python_version = "3.8"
